Favoritos a funcionar mas nao guardo o estado da pagina, mudança nos construtores do estados da pagina
