+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   http: "^0.11.3+16"
   async_loader: "^0.1.1"
   path_provider: "^0.4.1"
+  flutter_redux: "^0.5.1"
